Split `tree.js` to separate modules for models and views
